backend/pyproject.toml
[project]
name = "afcadinho-web-revamped"
version = "0.1.0"
description = ""
authors = [
    {name = "Afc_Adinho",email = "afcadinho@gmail.com"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "flask (>=3.1.3,<4.0.0)",
    "gunicorn (>=25.3.0,<26.0.0)",
    "flask-cors (>=6.0.2,<7.0.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "psycopg[binary] (>=3.3.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
